Group authentication types in `opt::auth`
