+++ lib/examples/transaction/main.rs
@@ -1,4 +1,4 @@
-use surrealdb::opt::Root;
+use surrealdb::opt::auth::Root;
 use surrealdb::sql::statements::BeginStatement;
 use surrealdb::sql::statements::CommitStatement;
